#3 put Segmentation Comparison in its own module
