+++ SegmentationComparison/Resources/UI/SegGroupsDialog.ui
